+++ nflow-engine/src/main/java/io/nflow/engine/config/db/H2DatabaseConfiguration.java
@@ -144,7 +144,7 @@ public class H2DatabaseConfiguration extends DatabaseConfiguration {
      * Returns SQL for a query with a limit of results.
      */
     @Override
-    public String limit(String query, String limit) {
+    public String limit(String query, long limit) {
       return query + " limit " + limit;
     }