+++ nflow-engine/src/main/java/io/nflow/engine/config/db/PgDatabaseConfiguration.java
@@ -110,7 +110,7 @@ public class PgDatabaseConfiguration extends DatabaseConfiguration {
      * Returns SQL for a query with a limit of results.
      */
     @Override
-    public String limit(String query, String limit) {
+    public String limit(String query, long limit) {
       return query + " limit " + limit;
     }