+++ nflow-engine/src/main/java/io/nflow/engine/internal/dao/WorkflowInstanceDao.java
@@ -548,7 +548,7 @@ public class WorkflowInstanceDao {
 
   private List<Integer> pollNextWorkflowInstanceIdsWithUpdateReturning(int batchSize) {
     String sql = updateInstanceForExecutionQuery() + " where id in ("
-        + sqlVariants.limit("select id from nflow_workflow " + whereConditionForInstanceUpdate(), Integer.toString(batchSize))
+        + sqlVariants.limit("select id from nflow_workflow " + whereConditionForInstanceUpdate(), batchSize)
         + ") and executor_id is null returning id";
     return jdbc.queryForList(sql, Integer.class);
   }
@@ -557,14 +557,9 @@ public class WorkflowInstanceDao {
     return transaction.execute(new TransactionCallback<List<Integer>>() {
       @Override
       public List<Integer> doInTransaction(TransactionStatus transactionStatus) {
-        String sql = sqlVariants.limit("select id, modified from nflow_workflow " + whereConditionForInstanceUpdate(),
-            Integer.toString(batchSize));
-        List<OptimisticLockKey> instances = jdbc.query(sql, new RowMapper<OptimisticLockKey>() {
-          @Override
-          public OptimisticLockKey mapRow(ResultSet rs, int rowNum) throws SQLException {
-            return new OptimisticLockKey(rs.getInt("id"), sqlVariants.getTimestamp(rs, "modified"));
-          }
-        });
+        String sql = sqlVariants.limit("select id, modified from nflow_workflow " + whereConditionForInstanceUpdate(), batchSize);
+        List<OptimisticLockKey> instances = jdbc.query(sql, (rs, rowNum) ->
+                new OptimisticLockKey(rs.getInt("id"), sqlVariants.getTimestamp(rs, "modified")));
         if (instances.isEmpty()) {
           return emptyList();
         }
@@ -582,7 +577,7 @@ public class WorkflowInstanceDao {
         boolean raceConditionDetected = false;
         for (OptimisticLockKey instance : instances) {
           int updated = jdbc.update(updateInstanceForExecutionQuery() + " where id = ? and modified = ? and executor_id is null",
-              instance.id, instance.modified);
+              instance.id, sqlVariants.tuneTimestampForDb(instance.modified));
           if (updated == 1) {
             ids.add(instance.id);
           } else {
@@ -598,7 +593,7 @@ public class WorkflowInstanceDao {
       private void updateNextWorkflowInstancesWithBatchUpdate(List<OptimisticLockKey> instances, List<Integer> ids) {
         List<Object[]> batchArgs = new ArrayList<>(instances.size());
         for (OptimisticLockKey instance : instances) {
-          batchArgs.add(new Object[] { instance.id, instance.modified });
+          batchArgs.add(new Object[] { instance.id, sqlVariants.tuneTimestampForDb(instance.modified) });
           ids.add(instance.id);
         }
         int[] updateStatuses = jdbc
@@ -686,8 +681,7 @@ public class WorkflowInstanceDao {
     conditions.add("w.executor_group = :executor_group");
     params.addValue("executor_group", executorInfo.getExecutorGroup());
     sql += " where " + collectionToDelimitedString(conditions, " and ") + " order by w.created desc";
-    sql = sqlVariants.limit(sql, ":limit");
-    params.addValue("limit", getMaxResults(query.maxResults));
+    sql = sqlVariants.limit(sql, getMaxResults(query.maxResults));
     List<WorkflowInstance> ret = namedJdbc.query(sql, params, new WorkflowInstanceRowMapper()).stream()
         .map(WorkflowInstance.Builder::build).collect(toList());
     for (WorkflowInstance instance : ret) {
@@ -712,11 +706,7 @@ public class WorkflowInstanceDao {
       public void processRow(ResultSet rs) throws SQLException {
         int parentActionId = rs.getInt(1);
         int childWorkflowInstanceId = rs.getInt(2);
-        List<Integer> children = instance.childWorkflows.get(parentActionId);
-        if (children == null) {
-          children = new ArrayList<>();
-          instance.childWorkflows.put(parentActionId, children);
-        }
+        List<Integer> children = instance.childWorkflows.computeIfAbsent(parentActionId, ArrayList::new);
         children.add(childWorkflowInstanceId);
       }
     }, instance.id);
@@ -732,8 +722,8 @@ public class WorkflowInstanceDao {
   private void fillActions(WorkflowInstance instance, boolean includeStateVariables, Long maxActions) {
     Map<Integer, Map<String, String>> actionStates = includeStateVariables ? fetchActionStateVariables(instance)
         : EMPTY_ACTION_STATE_MAP;
-    String limit = Long.toString(getMaxActions(maxActions));
-    String sql = sqlVariants.limit("select * from nflow_workflow_action where workflow_id = ? order by id desc", limit);
+    String sql = sqlVariants.limit("select nflow_workflow_action.* from nflow_workflow_action where workflow_id = ? order by id desc",
+            getMaxActions(maxActions));
     instance.actions.addAll(jdbc.query(sql, new WorkflowInstanceActionRowMapper(sqlVariants, actionStates), instance.id));
   }